+++ src/linking/generator.rs
@@ -220,27 +220,6 @@ pub fn get_bin_require_path(
 	luau_style_path(&path)
 }
 
-/// Generate a linking module for a script
-#[must_use]
-pub fn generate_script_linking_module(require_path: &str) -> String {
-	format!(r"return require({require_path})")
-}
-
-/// Get the require path for a script
-#[instrument(level = "trace", ret)]
-#[must_use]
-pub fn get_script_require_path(
-	base_dir: &Path,
-	script_file: &RelativePath,
-	destination_dir: &Path,
-) -> String {
-	let path = pathdiff::diff_paths(destination_dir, base_dir).unwrap();
-	tracing::debug!("diffed script path: {}", path.display());
-	let path = script_file.to_path(path);
-
-	luau_style_path(&path)
-}
-
 /// Errors for the linking module utilities
 pub mod errors {
 	use thiserror::Error;