+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Drop `build_files` ([RFC](https://github.com/pesde-pkg/pesde/issues/57)) by @daimond113
 - Remove pesde-managed script inheritance by @daimond113
+- Remove scripts packages by @daimond113
 
 ## [0.7.1] - 2025-08-24
 ### Fixed